tree: hide interface functions inside its file
The aim of this patch is to get rid of header files. We don't want to
duplicate function signatures in a header whenever we add an
interface. Instead of it, form a `UMT_STATE_ACTIONS` inside the
interface's file. This structure will be the ONLY way to interact with
the interface from outside its file.

As far as we did it, we can mark functions used in callbacks as static.